requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# OpenAI and LLM
openai==1.12.0
langchain==0.1.5
langchain-openai==0.0.5

# Twitter
twikit==2.3.3
websocket-client==1.7.0

# Web and API
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.1
PyYAML==6.0.1
pytz==2024.1
loguru==0.7.2
colorama==0.4.6

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
mypy==1.8.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
